changed SENT format - Fixes VPN pay function

+++ mobile-client-android/sentinel_android/app/src/main/java/sentinelgroup/io/sentinel/network/model/Session.java
@@ -13,7 +13,7 @@ public class Session implements Serializable {
     public long receivedBytes;
     @SerializedName("session_duration")
     public long sessionDuration;
-    public double amount;
+    public long amount;
     public long timestamp;
     @SerializedName("is_paid")
     public boolean isPaid;
